requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
beautifulsoup4==4.12.3
    # via -r requirements.in
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
distro==1.9.0
    # via openai
dnspython==2.6.1
    # via pymongo
fastapi==0.110.2
    # via -r requirements.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.5
    # via httpx
httptools==0.6.1
    # via uvicorn
httpx==0.27.0
    # via openai
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
openai==1.23.2
    # via -r requirements.in
pydantic==2.7.0
    # via
    #   -r requirements.in
    #   fastapi
    #   openai
pydantic-core==2.18.1
    # via pydantic
pymongo==4.6.3
    # via -r requirements.in
python-dotenv==1.0.1
    # via
    #   -r requirements.in
    #   uvicorn
pyyaml==6.0.1
    # via uvicorn
requests==2.31.0
    # via -r requirements.in
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
soupsieve==2.5
    # via beautifulsoup4
starlette==0.37.2
    # via fastapi
tqdm==4.66.2
    # via openai
typing-extensions==4.11.0
    # via
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.1
    # via requests
uvicorn[standard]==0.29.0
    # via -r requirements.in
uvloop==0.19.0
    # via uvicorn
watchfiles==0.21.0
    # via uvicorn
websockets==12.0
    # via uvicorn
